externalize basic result feature
